+++ lib/src/logic/android_config.g.dart
@@ -22,6 +22,12 @@ FirebaseAndroidConfig _$FirebaseAndroidConfigFromJson(
           : FirebaseAndroidNotification.fromJson(
               json['notification'] as Map<String, dynamic>),
       directBootOk: json['direct_boot_ok'] as bool?,
+      fcmOptions: json['fcm_options'] == null
+          ? null
+          : AndroidFcmOptions.fromJson(
+              json['fcm_options'] as Map<String, dynamic>),
+      bandwidthConstrainedOk: json['bandwidth_constrained_ok'] as bool?,
+      restrictedSatelliteOk: json['restricted_satellite_ok'] as bool?,
     );
 
 Map<String, dynamic> _$FirebaseAndroidConfigToJson(
@@ -34,9 +40,22 @@ Map<String, dynamic> _$FirebaseAndroidConfigToJson(
       'data': instance.data,
       'notification': instance.notification,
       'direct_boot_ok': instance.directBootOk,
+      'fcm_options': instance.fcmOptions,
+      'bandwidth_constrained_ok': instance.bandwidthConstrainedOk,
+      'restricted_satellite_ok': instance.restrictedSatelliteOk,
     };
 
 const _$AndroidMessagePriorityEnumMap = {
   AndroidMessagePriority.normal: 'NORMAL',
   AndroidMessagePriority.high: 'HIGH',
 };
+
+AndroidFcmOptions _$AndroidFcmOptionsFromJson(Map<String, dynamic> json) =>
+    AndroidFcmOptions(
+      analyticsLabel: json['analytics_label'] as String?,
+    );
+
+Map<String, dynamic> _$AndroidFcmOptionsToJson(AndroidFcmOptions instance) =>
+    <String, dynamic>{
+      'analytics_label': instance.analyticsLabel,
+    };