+++ lib/src/logic/fcm_options.g.dart
@@ -9,11 +9,9 @@ part of 'fcm_options.dart';
 FirebaseFcmOptions _$FirebaseFcmOptionsFromJson(Map<String, dynamic> json) =>
     FirebaseFcmOptions(
       analyticsLabel: json['analytics_label'] as String?,
-      image: json['image'] as String?,
     );
 
 Map<String, dynamic> _$FirebaseFcmOptionsToJson(FirebaseFcmOptions instance) =>
     <String, dynamic>{
       'analytics_label': instance.analyticsLabel,
-      'image': instance.image,
     };